Upd: minify javascript for release builds in master branch only

+++ build.sh
@@ -65,6 +65,12 @@ umask 0022
 VERSION=$(grep "  VERSION" CMakeLists.txt | sed 's/  VERSION //')
 COPYRIGHT="myMPD ${VERSION} | (c) 2018-2022 Juergen Mang <mail@jcgames.de> | SPDX-License-Identifier: GPL-3.0-or-later | https://github.com/jcorporation/mympd"
 
+MYMPD_MINIFY_JS="1"
+if [ -f .git/HEAD ] && ! grep -q "master" .git/HEAD
+then
+  MYMPD_MINIFY_JS="0"
+fi
+
 #check for command
 check_cmd() {
   for DEPENDENCY in "$@"
@@ -143,11 +149,16 @@ minify() {
   elif [ "$TYPE" = "js" ]
   then
     #shellcheck disable=SC2016
-    if ! perl -pe 's/^\s*//gm; s/^\s*\/?\*.*$//g; s/^\/\/.+$//g; s/^logDebug\(.*$//g; s/\/\*debug\*\/.*$//g; s/\s*$//gm;' "$SRC" > "${DST}.tmp"
+    if [ "$MYMPD_MINIFY_JS" = "0" ]
     then
-      rm -f "${DST}.tmp"
-      echo_error "Error minifying $SRC"
-      exit 1
+      cp "$SRC" "${DST}.tmp"
+    else
+      if ! perl -pe 's/^\s*//gm; s/^\s*\/?\*.*$//g; s/^\/\/.+$//g; s/^logDebug\(.*$//g; s/\/\*debug\*\/.*$//g; s/\s*$//gm;' "$SRC" > "${DST}.tmp"
+      then
+        rm -f "${DST}.tmp"
+        echo_error "Error minifying $SRC"
+        exit 1
+      fi
     fi
   elif [ "$TYPE" = "json" ]
   then
@@ -215,7 +226,12 @@ createassets() {
 
   echo "Combining and compressing javascript"
   echo "//${COPYRIGHT}" > "$MYMPD_BUILDDIR/htdocs/js/copyright.min.js"
-  JSFILES="dist/bootstrap-native/bootstrap-native.min.js dist/long-press-event/long-press-event.min.js"
+  if [ "$MYMPD_MINIFY_JS" = "0" ]
+  then
+    JSFILES="dist/bootstrap-native/bootstrap-native.js dist/long-press-event/long-press-event.js"
+  else
+    JSFILES="dist/bootstrap-native/bootstrap-native.min.js dist/long-press-event/long-press-event.min.js"
+  fi
   JSFILES="$JSFILES $MYMPD_BUILDDIR/htdocs/js/*.min.js"
   for F in $JSFILES
   do
